restructure providers directory
