+++ packages/react-storage/src/components/StorageBrowser/actions/configs/context.ts
@@ -1,6 +1,10 @@
 import { createContextUtilities } from '@aws-amplify/ui-react-core';
 import { ActionConfigs } from './types';
 
+export interface ActionConfigsProviderProps {
+  actions?: ActionConfigs;
+}
+
 const defaultValue: ActionConfigs = {};
 
 export const { useActionConfigs, ActionConfigsProvider } =
