+++ packages/react-storage/src/components/StorageBrowser/actions/index.ts
@@ -1,4 +1,11 @@
 export {
+  ActionConfigs,
+  ActionConfigsProvider,
+  ActionConfigsProviderProps,
+  useActionConfigs,
+} from './configs';
+
+export {
   createFolderHandler,
   CreateFolderHandler,
   listLocationItemsHandler,
