+++ packages/react-storage/src/components/StorageBrowser/providers/store/index.ts
@@ -1,0 +1,1 @@
+export { StoreProvider, StoreProviderProps } from './StoreProvider';
